requirements.txt
Django>=4.2.0,<5.0
Pillow>=9.0.0
django-crispy-forms>=2.0
crispy-tailwind>=0.5.0
python-decouple>=3.6
whitenoise>=6.0
